Updated deploy/debug scripts

+++ .vscode/generate_launch.sh
@@ -1,13 +1,13 @@
 #!/bin/bash
 
-#get the direcotry of this script
+#get the directory of this script
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 target_device_file=$DIR/target.device
 launch_json_file=$DIR/launch.json
 
 
-#check argument 1 if its present set default_ip to it  and write it to target.device
+#check argument 1 if it's present set default_ip to it and write it to target.device
 if [ -n "$1" ]; then
     default_ip=$1
     echo $default_ip > $target_device_file
@@ -24,12 +24,9 @@ else
     ip_address=$default_ip
 fi
 
-# Construct the launch.json content with the updated IP address
-launch_json_content=$(cat << EOF
+#create configuration 
+ltpp3g2_launch_config=$(cat << EOF
 {
-    "version": "0.2.0",
-    "configurations": [
-        {
             "name": "Attach to LTPP3G2",
             "type": "cppdbg",
             "request": "launch",
@@ -51,11 +48,25 @@ launch_json_content=$(cat << EOF
             "miDebuggerPath": "/usr/bin/gdb-multiarch",
             "targetArchitecture": "arm",
             "preLaunchTask": "LTPP3G2: Deploy and Launch Debug"
-        }
+}
+EOF
+)
+
+# Construct the launch.json content with the updated IP address
+launch_json_content=$(cat << EOF
+{
+    "version": "0.2.0",
+    "configurations": [
+        $ltpp3g2_launch_config
     ]
 }
 EOF
 )
 
-# Save the launch.json content to a file
-echo "$launch_json_content" >  $launch_json_file
+#if launch.json does not exist create it
+if [ ! -f "$launch_json_file" ]; then
+     echo "$launch_json_content" >  $launch_json_file
+else 
+#use JQ to replace from the configurations array the confiuration named "Attach to LTPP3G2" with the one in ltpp3g2_launch_config variable
+    jq --argjson ltpp3g2_launch_config "$ltpp3g2_launch_config" '.configurations |= map(if .name == "Attach to LTPP3G2" then $ltpp3g2_launch_config else . end)' $launch_json_file > $launch_json_file.tmp && mv $launch_json_file.tmp $launch_json_file
+fi
